requirements.txt
google-cloud-firestore==2.11.1
fireo==2.2.1
starlette==0.37.2
Jinja2==3.1.4
python-multipart==0.0.9
gunicorn==22.0.0
itsdangerous==2.2.0
flask==3.0.3
requests==2.31.0
python-dateutil==2.9.0
Flask-HTTPAuth==4.8.0
cryptography==42.0.7
pytz==2024.1